+++ include/layered_hardware_gazebo/posvel_pid_mode.hpp
@@ -12,6 +12,8 @@
 #include <transmission_interface/transmission_interface_loader.h> //for RawJointData
 #include <urdf/model.h>
 
+#include <gazebo/physics/physics.hh>
+
 #include <boost/algorithm/clamp.hpp>
 #include <boost/math/special_functions/fpclassify.hpp> // for isnan()
 
@@ -19,8 +21,8 @@ namespace layered_hardware_gazebo {
 
 class PosVelPIDMode : public OperationModeBase {
 public:
-  PosVelPIDMode(const urdf::Joint &desc)
-      : OperationModeBase("posvel_pid"),
+  PosVelPIDMode(const urdf::Joint &desc, const gzp::JointPtr joint)
+      : OperationModeBase("posvel_pid", joint),
         eff_lim_(desc.limits ? std::abs(desc.limits->effort) : 1e10) {}
 
   virtual ~PosVelPIDMode() {}