+++ src/layered_hardware_gazebo_plugins.cpp
@@ -1,5 +1,6 @@
-#include <layered_hardware/layer_base.hpp>
 #include <layered_hardware_gazebo/gazebo_joint_layer.hpp>
+#include <layered_hardware_gazebo/gazebo_layer_base.hpp>
 #include <pluginlib/class_list_macros.h>
 
-PLUGINLIB_EXPORT_CLASS(layered_hardware_gazebo::GazeboJointLayer, layered_hardware::LayerBase);
+PLUGINLIB_EXPORT_CLASS(layered_hardware_gazebo::GazeboJointLayer,
+                       layered_hardware_gazebo::GazeboLayerBase);