+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package layered_hardware_gazebo
 0.0.4 (next version)
 --------------------
 * Support custom control frequency, independent from the simulation step
+* Support user-defined gazebo layer plugins which base type is GazeboLayerBase
 
 0.0.3 (2020-04-30)
 ------------------